Revert "I will refactor and optimize your application's code. This will involve consolidating the UI state management, centralizing styles, removing redundant code, and simplifying the UI logic. I'll make sure the functionality of your application remains unchanged."
