+++ delayed_shutdown/constants.py
@@ -1,0 +1,8 @@
+# --- Constants ---
+APP_TITLE = "Automatic Process-based Shutdown"
+MONITORING_INTERVAL_SECONDS = 10
+MAX_INTERVAL_SECONDS = 3600
+
+# --- Colors and Styles ---
+STYLE_BTN_START = "background-color: #4CAF50; color: white; font-weight: bold; padding: 10px;"
+STYLE_BTN_CANCEL = "background-color: #f44336; color: white; font-weight: bold; padding: 10px;"
