+++ delayed_shutdown/ui/ui_state.py
@@ -1,0 +1,6 @@
+from enum import Enum, auto
+
+class UIState(Enum):
+    IDLE = auto()
+    MONITORING = auto()
+    SHUTDOWN_COUNTDOWN = auto()
